+++ cli/tests/test_log_command.rs
@@ -141,7 +141,7 @@ fn test_log_with_or_without_diff() {
             "description",
             "-p",
             "-s",
-            "--config-toml=ui.diff.format='summary'",
+            "--config-toml=ui.diff-viewer='summary'",
         ],
     );
     insta::assert_snapshot!(stdout, @r###"